Revert "shill: Convert code to use the newest version of libchrome."
This reverts commit ffebe0c489e0ecccf497547d7078c30cbc8072e2

Unit tests ran on local machine, but wouldn't run on CQ because of some kind of environment difference.

Reviewed-on: https://gerrit.chromium.org/gerrit/17969
Tested-by: Eric Shienbrood <ers@chromium.org>

+++ shill/ip_address.cc
@@ -54,7 +54,7 @@ size_t IPAddress::GetAddressLength(Family family) {
 }
 
 // static
-size_t IPAddress::GetPrefixLengthFromMask(Family family, const string &mask) {
+int IPAddress::GetPrefixLengthFromMask(Family family, const string &mask) {
   switch (family) {
     case kFamilyIPv4: {
       in_addr_t mask_val = inet_network(mask.c_str());
